feishu-alert-service/requirements.txt
lark-oapi>=1.5.3,<2
openai>=2.21.0,<3
pyyaml>=6.0.2,<7
mcp>=1.2.0,<2
